requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
